requirements.txt
ccxt>=4.3.0
pandas>=2.2.2
numpy>=1.26.4
requests>=2.32.3
python-dotenv>=1.0.1
